Skiller lydkanal i flere filer
