+++ core/src/apu/length_timer.rs
@@ -1,0 +1,4 @@
+#[derive(Default)]
+pub struct LengthTimer {
+    pub enabled: bool,
+}
